% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialstock.R
\name{materialstock_updateUI}
\alias{materialstock_updateUI}
\title{materialstock_update生成器界面}
\usage{
materialstock_updateUI(
  tabTitle = "天铭数据中台",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = materialstock_updateUI_left,
  func_right = materialstock_updateUI_right,
  func_bottom = materialstock_updateUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
materialstock_update生成器界面
}
\examples{
materialstock_updateUI()
}
