% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialstock.R
\name{materialstock_updateUI_right}
\alias{materialstock_updateUI_right}
\title{生成区域}
\usage{
materialstock_updateUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
materialstock_updateUI_bottom()
}
